+++ packages/polyfills/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Breaking change
+
+- Removed the deprecated dependency `@babel/polyfill` and move to `core-js@^3.0.0`.
+- Removed the `@shopify/polyfill/url` polyfill, it's automatically provided with `core-js`.
+- Removed the `@shopify/polyfill/base` polyfill, polyfills will be automatically provided base on usage.
+- Removed the `@shopify/polyfill/unhandled-rejection` polyfill, polyfills will be automatically provided base on usage.
+
+**Note**: This new version needs at least `sewing-kit@^0.150.0`.
+
 ## [1.3.2] - 2021-03-03
 
 ### Fixed