+++ packages/polyfills/CHANGELOG.md
@@ -7,6 +7,8 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## Unreleased -->
 
+## 2.0.1 - 2021-05-28
+
 ### Fixed
 
 - Prevent this package from being polyfilled. [#1921](https://github.com/Shopify/quilt/pull/1921)